-- Global File creation for the execution of global application. 	new file:   __init__.py 	new file:   __main__.py 	new file:   apis/__init__.py
As required for the global execution, the following operation was made:
	renamed:    app.py -> app_example.py
	created:    app.py

-- Database

Following the 'scaling your project'(https://flask-restx.readthedocs.io/en/latest/scaling.html)
the database files were move to apis/

apis/database/app.py was modified to support the proper json

proper testing is yet to be implemented

-- User

apis/user/app.py was created and some methods declared, how ever not implemented.
Things which have to be modified are:
	- post: verification on arguments, such as valid age, valid password, etc.
	- get: token is created, but isn't stored anywhere yet.
	- update: initially was thought to be the method to modify the password. However, I'm considering to change it to another subaddress

The Hist class can be completely ignored for the meantime. This would be the class related
to the address to get the user buy history. However, further analysis shall be made, as I get a
better comprehension of the libraries.

+++ src/apis/database/scheme.py
@@ -3,7 +3,7 @@
 
 from datetime import date, datetime
 
-from sqlalchemy import Column, Date, DateTime, ForeignKey, Integer, Float, String
+from sqlalchemy import Column, Date, DateTime, ForeignKey, Integer, Float, String, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from sqlalchemy import create_engine
@@ -27,7 +27,7 @@ class User(Base, Serializable):
 	birthday = Column(Date)
 	email = Column(String(255), unique=True)
 	passwd = Column(String(255))
-	phone = Column(Integer, unique=True)
+	phone = Column(BigInteger, unique=True)
 
 	def __repr__(self):
 		return f"User: {self.name}"
